requirements.txt
flask[async]==3.0.2
motor==3.3.2
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-asyncpg==0.43b0
opentelemetry-instrumentation-flask==0.43b0
opentelemetry-instrumentation-pymongo==0.43b0
asyncpg==0.29.0
opentelemetry-exporter-otlp-proto-http==1.22.0